Create an interface to launch the app
